--- a/db/migrate/20250922200515_add_index_to_user_id_on_invite_table.rb
+++ b/db/migrate/20250922200515_add_index_to_user_id_on_invite_table.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class AddIndexToUserIdOnInviteTable < ActiveRecord::Migration[7.1]
+  disable_ddl_transaction!
+
+  def change
+    add_index :invites, :user_id, name: 'index_invites_on_user_id', algorithm: :concurrently
+  end
+end