--- a/db/schema.rb
+++ b/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.1].define(version: 2025_09_10_074841) do
+ActiveRecord::Schema[7.1].define(version: 2025_09_22_200515) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_buffercache"
   enable_extension "pg_stat_statements"
@@ -876,6 +876,7 @@ ActiveRecord::Schema[7.1].define(version: 2025_09_10_074841) do
     t.integer "user_id"
     t.index ["dossier_id"], name: "index_invites_on_dossier_id"
     t.index ["email", "dossier_id"], name: "index_invites_on_email_and_dossier_id", unique: true
+    t.index ["user_id"], name: "index_invites_on_user_id"
   end
 
   create_table "labels", force: :cascade do |t|