[Selon skylight](https://www.skylight.io/app/applications/zAvWTaqO0mu1/recent/6h/endpoints/Users::DossiersController%23index?responseType=html) , ce endpoint est le second en terme d'agonie. 
Une  majorité du temps est passé en sql, plus précisément sur la méthode `all_dossier_procedures.distinct(:procedure_id).order(:libelle).pluck(:libelle, :id)`

<img width="1896" height="428" alt="Screenshot 2025-09-22 at 23-10-51 _dossiers (1279 3 ms) - Profiling Results" src="https://github.com/user-attachments/assets/7373eedf-00db-424b-babf-b1eebfc0055c" />

la première étape consiste à mettre un index sur la table invité pour accélérer `current_user.dossiers_invites.visible_by_user` (70ms -> 1ms)

<img width="1894" height="428" alt="Screenshot 2025-09-22 at 23-14-41 _dossiers (513 3 ms) - Profiling Results" src="https://github.com/user-attachments/assets/85b8dc80-07c3-46fe-b103-ef7dd4ccf6b7" />

puis élagage du sql `current_user.dossiers_invites.visible_by_user == current_user.dossiers_invites`

Finalement, le `or` est étonnant pas le top du top en terme d’optimisation (et rails étant pas d'accord avec ma syntaxe), on utilise (merci claude) l'union que [postgresql préfère](https://dba.stackexchange.com/questions/293836/why-is-an-or-statement-slower-than-union)

<img width="1892" height="380" alt="Screenshot 2025-09-22 at 23-21-15 _dossiers (465 7 ms) - Profiling Results" src="https://github.com/user-attachments/assets/e2ebbbca-0721-4635-982d-ed35b2804f1e" />