@LeSim 2025-09-24
> super clair, merci pour les détails.
> 
>     * est-ce que cela ne vaudrait pas le coup de faire du user_id une référence / vraie FK ?
> 
>     * légèrement surpris que rails laisse passer un where avec une interpolation sans warning plutôt que passer par des `?`. Ou alors il est assez malin pour voir que ça provient de strings de confiance ? (mais j'en doute )

Bien vu pour la fk. J'ai essayé mais bien sur ... notre donnée est pourrie, du coup je propose de faire ca dans une autre pr  avec la maintenance task qui va bien (mais peut etre pas tout de suite), histoire de profiter du gain de cette pr.

Pour la partie injection, brakeman reconstruit la query complete et regarde s'il y a un params dedans. Par exemple, je peux le faire raler en replacant `user_revision_ids = current_user.dossiers.visible_by_user.select(:revision_id).to_sql` par `params[:revision_id]`

```
Confidence: Medium
Category: SQL Injection
Check: SQL
Message: Possible SQL injection
Code: ProcedureRevision.where("id IN (#{"(#{params[:revision_id]}) UNION (#{current_user.dossiers_invites.select(:revision_id).to_sql})"})")
File: app/controllers/users/dossiers_controller.rb
```